Fix shared memory

+++ Sources/CameraSharedMem/CameraSharedMem.c
@@ -9,6 +9,6 @@
 #include <sys/mman.h>
 #include "CameraSharedMem.h"
 
-int shmOpen(const char *name, int flags, int mode) {
+int shmOpen(const char *name, int flags, unsigned short int mode) {
     return shm_open(name, flags, mode);
 }
